feat(Versioning): Follow Inertia.js spec on versioning
Allows a version setting to be configured by the user, usually
pointing at a hash of the Webpack manifest file.  If the version
does not match the incoming version header, then a 409 Conflict
will be returned.  Inertia.js will then do a full page refresh
to the value of the `X-Inertia-Location` header to get the latest
asset changes.